[FIX] 년차 컬럼명 수정 및 테이블 생성 확인

+++ Harbor_employee/src/main/java/com/example/harbor_employee/domain/Eworks.java
@@ -22,7 +22,7 @@ import javax.persistence.*;
 public class Eworks {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer payId;
+    private Long payId;
 
     @Column(nullable = false, length = 10)
     private String payStatusCode;
